requirements.txt
fastapi
uvicorn[standard]
transformers
torch
torchvision
torchaudio
pillow
numpy
pandas